refactor the post processing code add confidence interval
